\name{phyloRcppExamples-package}
\alias{phyloRcppExamples-package}
\alias{phyloRcppExamples}
\docType{package}
\title{Examples of using Rcpp for phylogenetic inference
}
\description{
This package provides C++ code and some simple R wrapper functions that
demonstrate how Rcpp package can be used for certain computational tasks
common in statistical phylogenetics.
}
\details{
\tabular{ll}{
Package: \tab phyloRcppExamples\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-10\cr
License: \tab Some sort of GPL\cr
LazyLoad: \tab yes\cr
}
See examples at http://markovjumps.blogspot.com/
}
\author{
Vladimir N. Minin 

Maintainer: Vladimir N. Minin<vminin@uw.edu>
}
\references{
To be added.
}
\keyword{ package }
